flake: remove, replace dependency management with npins (#3)
Pulling the flake inputs out for use in non-flake evaluations is a
little hairy. We're already intentionally trying to avoid relying on the
flake as much as possible, and the only reason it's there is so that
flake users can override nixpkgs with the usual `follows` mechanism if
they wanted. There's other ways to achieve the same thing without
relying on `follows` though, so we don't even need flakes for that.

This PR drops the flakes, and replaces the dependency management with
[npins](https://github.com/andir/npins), which is much nicer to use from
non-flake code.